refactor: encapsulated SSE.js functions into new SSEService class
